data_hub/backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
psycopg2-binary==2.9.9

# ── Validation ────────────────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# ── HTTP client (connection testing) ─────────────────────────────────────────
httpx==0.27.2

# ── Utilities ─────────────────────────────────────────────────────────────────
python-multipart==0.0.12
python-dotenv==1.0.1
structlog==24.4.0

# ── Data processing ───────────────────────────────────────────────────────────
pandas==2.2.3
